+++ test/test_get_latest_tag.py
@@ -1,7 +1,5 @@
-import pytest
-import os
 from steinnews import get_latest_tag
-from steinnews.exceptions import NoChanges
+
 
 def test_get_latest_tag(file_in, expected_version):
     with open(file_in, "r") as file: