+++ test/test_validate_latest_content.py
@@ -1,0 +1,14 @@
+# from steinnews import validate_latest_content
+
+# def test_valid():
+#     assert validate_latest_content("""Latest
+
+# -+
+
+# * Minor: test
+
+# * Patch: test
+
+# * Major: test
+
+# """) == None